Added support for a malloc-based version of pybloomfilter's BloomFilter
Basically, for SEOmoz's purposes, we didn't need any of the features that mmap
would support: persistence, the ability to exceed physical memory without swap,
etc. The problem we were running into was that we instead needed to make /a lot/
of bloom filters and we were finding it irritating to have to up our open file
descriptor limit. After all, we just wanted to make a bunch of bloom filters that
would fit in memory.